+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The project adheres to the [Haskell Package Versioning Policy](https://pvp.haske
 
 ## Unreleased
 
+- Awesome upcoming features!
+
+## 0.0.1.1 - Small cosmetic improvements - 2025-03-28
+
 - Demonstration:
   - Added menu and option to change wled device URL via terminal or command line argument
   - Added tool to traverse every single LED (to be able to better detect positions)